deduplicate some bsd tuntap code, still untested
